Refactor/cleanup
* Moved data logic into the data-fetcher service
* Partial progress on #3
