+++ src/app/app.component.html
@@ -2,6 +2,9 @@
     <!-- Toolbar -->
     <mat-toolbar color='primary' class='toolbar'>
         <button mat-icon-button (click)="countryGraph.toggle()"><mat-icon>menu</mat-icon></button>
+
+        <span class="spacer"></span>
+
         <h1>Covid data browser</h1>
 
         <span class="spacer"></span>
